refactor(email): use first-class templates
