--- components/emails/variants/legal-document-notice-information.tsx
@@ -1,16 +1,0 @@
-import type { AppLocale } from "@/i18n/locale-registry";
-
-import { renderEmailPreview } from "../preview-manifest";
-
-type Props = { locale?: AppLocale };
-
-export default function LegalDocumentNoticeInformationPreview({ locale }: Props) {
-  return renderEmailPreview("legal-document-notice", {
-    locale,
-    variant: "information",
-  });
-}
-
-LegalDocumentNoticeInformationPreview.PreviewProps = {
-  locale: "en",
-} satisfies Props;
