+++ ee/lifecycle/__tests__/send-legal-document-notices.interactor.test.ts
@@ -11,6 +11,13 @@ const mockLegalDocumentNotice = vi.hoisted(() =>
     type: "legal-document-notice",
   })),
 );
+const mockLegalDocumentNoticeInformation = vi.hoisted(() =>
+  vi.fn((props: unknown) => ({
+    key: null,
+    props,
+    type: "legal-document-notice-information",
+  })),
+);
 const mockSupplierDeadline = vi.hoisted(() => ({
   value: null as string | null,
 }));
@@ -25,8 +32,11 @@ vi.mock("@/constants/legal-documents", async (importOriginal) => {
     },
   };
 });
-vi.mock("@/components/emails/legal-document-notice", () => ({
+vi.mock("@/components/emails/legal-document-notice-contract", () => ({
   default: mockLegalDocumentNotice,
+}));
+vi.mock("@/components/emails/legal-document-notice-information", () => ({
+  default: mockLegalDocumentNoticeInformation,
 }));
 vi.mock("@/core/decorators/system-interactor.decorator", () => ({
   SystemInteractor: (target: unknown) => target,
@@ -273,6 +283,7 @@ describe("SendLegalDocumentNoticesInteractor", () => {
 
     emailService.send.mockReset().mockResolvedValue(undefined);
     mockLegalDocumentNotice.mockClear();
+    mockLegalDocumentNoticeInformation.mockClear();
     await invoke();
 
     expect(emailService.send).toHaveBeenCalledTimes(1);
@@ -306,6 +317,11 @@ describe("SendLegalDocumentNoticesInteractor", () => {
       "Subprocessors",
     ]);
     expect(documentNames(emailService.send.mock.calls[2][0])).toEqual(["Privacy Policy"]);
+    expect(emailService.send.mock.calls.map(([email]) => email.react.type)).toEqual([
+      "legal-document-notice",
+      "legal-document-notice",
+      "legal-document-notice-information",
+    ]);
 
     expect(eventService.publish).toHaveBeenCalledTimes(3);
     for (const [index, expectedEntityId] of ["admin-1", "admin-2", "member-1"].entries()) {
